% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorplot-package.R
\docType{package}
\name{factorplot-package}
\alias{factorplot-package}
\title{factorplot}
\description{
Factorplot is a way to summarize and plot information from categorical
predictors from linear models and GLMs.  It creates all simple contrasts and
analytical standard errors for those contrasts.
}
\details{
\tabular{ll}{ Package: \tab factorplot\cr Type: \tab Package\cr Version:
\tab 1.2.3\cr Date: \tab 2024-05-15\cr License: \tab GPL (>=2)\cr LazyLoad:
\tab yes\cr } After a linear model or GLM has been estimated, the factorplot
command creates all pairwise differences among the levels (including the
reference category) of the indicated factor as well as their associated
standard errors to facilitate hypothesis testing directly.  The print method
prints the pairwise difference, standard error, p-value and
Bonferroni-corrected p-value.  The summary method prints the number of
significant positive/negative pairwise differences.  The plot method makes
something akin to an upper-triangular levelplot that indicates whether
differences are positive/negative and statistically significant.
}
\references{
Armstrong, David A., II. 2013. factorplot: Improving
Presentation of Simple Contrasts in Generalized Linear Models.  \emph{The R
Journal} \bold{5(2)}: 4--15.
}
\author{
Dave Armstrong Maintainer: Dave Armstrong <dave@quantoid.net>
}
\keyword{internal}
\keyword{package}
