% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorplot.R
\name{squares}
\alias{squares}
\title{Auxiliary Function to Plot a Square}
\usage{
squares(ll, width = 1, col)
}
\arguments{
\item{ll}{The (\code{x},\code{y}) coordinate of the lower-left corder of the
square}

\item{width}{a scalar indicating how wide the squares should be}

\item{col}{a color with which the square will be filled in}
}
\value{
\item{square}{A square is printed on the graph, but nothing else is
returned}
}
\description{
An auxiliary function to plot squares, used by the
\code{\link{plot.factorplot}} function
}
\details{
This is a function called by \code{\link{plot.factorplot}} and not intended
to be directly used by the user; however, it is possible that this could be
of more general use as a utility.  The function is simply a wrapper to
\code{polygon} that obviates the need to specify all (\code{x},\code{y})
coordinates for the polygon.
}
\author{
Dave Armstrong
}
